requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-session==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
requests==2.26.0
python-dotenv==0.19.0
werkzeug==2.0.1
google-generativeai==0.3.0
aiohttp==3.8.1
yfinance==0.2.36
pandas==2.1.0
numpy==1.24.0
textblob==0.17.1
newsapi-python==0.2.7
flask-caching 
bson
pymongo
bcrypt